requirements.txt
Flask==2.2.0
Flask-Session==0.5.0
Flask-SeaSurf==1.1.1
Flask-Talisman==1.1.0
psycopg2-binary==2.9.11
Werkzeug==2.2.2
Waitress==3.0.0
python-dotenv==1.0.0
Pillow>=10.0.0
python-magic-bin==0.4.14
bcrypt==4.1.2
Flask-Limiter==2.0.1
